requirements.txt
azure-identity==1.25.1
agent-framework==1.0.0b251211
agent-framework-a2a==1.0.0b251211
agent-framework-ag-ui==1.0.0b251211
agent-framework-azure-ai==1.0.0b251211
agent-framework-chatkit==1.0.0b251211
agent-framework-core==1.0.0b251211
agent-framework-devui==1.0.0b251211
azure-ai-projects==2.0.0.b2
azure-ai-evaluation==1.13.7
azure-search-documents==11.6.0
azure-monitor-opentelemetry-exporter==1.0.0b46
openai==2.8.0
pydantic==2.12.4
aiohttp==3.13.2
aiohttp-sse==2.2.0
gunicorn==23.0.0
python-dotenv==1.2.1